refactor(verification): extract unified verification module
